Synthetic board class

+++ OpenBCI_GUI/Board.pde
@@ -15,7 +15,7 @@ interface Board {
     
     public int getNumChannels();
 
-    public int[] getLastAccelValues();
+    public float[] getLastAccelValues();
 
     public void setChannelActive(int channelIndex, boolean active);
 };
